pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "indexly"
version = "1.0.6"
description = "A local file indexing and search tool with FTS5, regex, and CSV analysis support."
readme = { file = "src/indexly/docs/README.md", content-type = "text/markdown" }

# Optional but strongly recommended
# Long PyPI-facing description for SEO 
long_description = """
Indexly is a fast, local-first file indexing and search engine powered by SQLite FTS5.
It scans documents, media, archives, spreadsheets, images, and structured data, then
provides instant full-text search, regex matching, metadata extraction, CSV profiling,
time-series visualizations, and XML/YAML/JSON parsing.

### Key Features
- **FTS5 full-text search** (fast, incremental, cached)
- **Regex search engine** with snippet extraction
- **OCR support** using Tesseract for image/PDF text extraction
- **CSV & JSON analysis** with summary statistics and pretty printed tables
- **XML + YAML loader** with validation and tree-view rendering
- **Document parsing**: PDF, DOCX, PPTX, XLSX, EPUB, ODT, MSG, EML
- **Image metadata extraction** (EXIF, dimensions, camera, timestamps)
- **Visualization**: ASCII plots, matplotlib charts, Plotly web graphs
- **Async real-time folder watching** with automatic re-indexing
- **Tagging system** for organizing files
- **Rich CLI output** powered by the `rich` library
- **Cross-platform**: Linux, macOS, Windows

### Why Indexly?
Indexly is built for developers, analysts, and power users who want a transparent,
lightweight alternative to desktop search tools—without sending data to the cloud.

### Links
- Documentation: https://projectindexly.com
- Source Code: https://github.com/kimsgent/project-indexly
- Issues: https://github.com/kimsgent/project-indexly/issues
"""

license = { text = "MIT" }
authors = [
    { name = "N. K Franklin-Gent" },
    { name = "OpenAI ChatGPT", email = "support@openai.com" }
]
requires-python = ">=3.8"
keywords = ["indexing", "search", "fts5", "regex", "csv", "asyncio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
    "Topic :: Database",
    "Topic :: Text Processing :: Indexing"
]

dependencies = [
    # Core NLP, document, and indexing dependencies
    "nltk>=3.8.0",
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.13",
    "Pillow>=11.3.0",
    "python-docx>=1.2.0",
    "openpyxl>=3.1.5",
    "rapidfuzz>=2.16.0",
    "fpdf2>=2.8.3",
    "reportlab>=4.4.2",
    "beautifulsoup4>=4.12.0",
    "extract_msg>=0.25.0",
    "eml-parser>=1.0.7",
    "PyPDF2>=3.0.1",
    "watchdog>=3.0.0",
    "colorama>=0.4.6",
    "python-pptx>=0.6.21",
    "ebooklib>=0.18",
    "odfpy>=1.4.1",
    "rich>=13.7.1",
    "pandas>=2.1.1",
    "PyYAML>=6.0.0",
    "xmltodict>=0.13.0",
    "requests>=2.31.0",

    # --- Visualization & Analysis Support ---
    # Terminal-based ASCII plots
    "plotext>=5.2.8",
    # Static plotting
    "matplotlib>=3.8.0",
    # Interactive web-based visualizations
    "plotly>=5.22.0",
    # Numeric & scientific utilities
    "numpy>=1.26.0",
    # Optional stats & data profiling tools (recommended for analyze-csv)
    "scipy>=1.12.0",
    "statsmodels>=0.14.0",
]

[tool.hatch.envs.dev]
dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "flake8>=6.1.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.6.0"
]

[optional-dependencies]
exiftool = []

[project.urls]
Homepage = "https://projectindexly.com"
Documentation = "https://projectindexly.com"
Source = "https://github.com/kimsgent/project-indexly"
Issues = "https://github.com/kimsgent/project-indexly/issues"

# ✅ FIXED BUILD CONFIG
[tool.hatch.build]
packages = ["src/indexly"]


[tool.hatch.build.targets.sdist]
include = [
  "LICENSE.txt",
  "README.md",
  "src/indexly/docs/README.md",
  "src/indexly/assets/fonts/*.ttf",
  "src/indexly/assets/csv/*.csv",
  "src/indexly/ignore_defaults/presets/**",
]

[tool.hatch.build.targets.wheel]
include = [
  "LICENSE.txt",
  "src/indexly/LICENSE.txt",
  "src/indexly/docs/README.md",
  "src/indexly/assets/fonts/*.ttf",
  "src/indexly/assets/csv/*.csv",
  "src/indexly/ignore_defaults/presets/**",
]

[project.scripts]
indexly = "indexly.__main__:main"
